+++ include/mvh/interrupt.h
@@ -1,0 +1,8 @@
+#ifndef MVH_INTERRUPT_H
+#define MVH_INTERRUPT_H
+
+void interrupt_init(void);
+void interrupt_enable(void);
+void interrupt_disable(void);
+
+#endif